backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
tushare==1.2.89
anthropic==0.28.0
pydantic==2.7.4
python-dotenv==1.0.1
